Reorganize conditions
