+++ wpilib/src/command/mod.rs
@@ -2,9 +2,12 @@
 pub mod manager;
 pub mod command_hid;
 pub mod commands;
+pub mod conditions;
 #[cfg(test)]
 mod test;
 
 pub use commands::Command;
 pub use manager::CommandManager;
 pub use manager::ConditionalScheduler;
+pub use conditions::OnTrue;
+pub use conditions::WhileTrue;
